refactor(LC037): split raw-sql construction analysis helpers
- separate construction detection and local-resolution logic
- keep analyzer behavior unchanged while reducing local complexity
