+++ src/LinqContraband/Analyzers/RawSqlAndSecurity/LC037_RawSqlStringConstruction/RawSqlStringConstructionLocalResolution.cs
@@ -1,0 +1,41 @@
+using LinqContraband.Extensions;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Operations;
+
+namespace LinqContraband.Analyzers.LC037_RawSqlStringConstruction;
+
+public sealed partial class RawSqlStringConstructionAnalyzer
+{
+    private static bool TryResolveLocalValue(ILocalSymbol local, IOperation? executableRoot, out IOperation value)
+    {
+        value = null!;
+
+        if (executableRoot == null)
+            return false;
+
+        foreach (var descendant in executableRoot.Descendants())
+        {
+            if (descendant is IVariableDeclarationOperation declaration)
+            {
+                foreach (var declarator in declaration.Declarators)
+                {
+                    if (!SymbolEqualityComparer.Default.Equals(declarator.Symbol, local) || declarator.Initializer == null)
+                        continue;
+
+                    value = declarator.Initializer.Value;
+                    return true;
+                }
+            }
+
+            if (descendant is ISimpleAssignmentOperation assignment &&
+                assignment.Target.UnwrapConversions() is ILocalReferenceOperation targetLocal &&
+                SymbolEqualityComparer.Default.Equals(targetLocal.Local, local))
+            {
+                value = assignment.Value;
+                return true;
+            }
+        }
+
+        return false;
+    }
+}
